Better cleanup for BuggyPackageDriver.xml file.

+++ Tests/SqlActorZoneLike/SqlActorPackage.Tests/BuggyPackageTests.cs
@@ -42,32 +42,39 @@ namespace SqlActorPackage.Tests
         [Test]
         public void failing_db_setup_does_not_execute_SettleContent()
         {
-            using( TestHelper.Monitor.OpenInfo( "failing_db_setup_does_not_execute_SettleContent (1/2)" ) )
+            try
             {
-                File.WriteAllText( _configFile, @"<Error ErrorStep=""Install"" />" );
-                SharedEngine.Reset();
-                SharedEngine.EngineResult.Status.Should().Be( RunStatus.Failed );
-                using( var ctx = new SqlStandardCallContext( TestHelper.Monitor ) )
+                using( TestHelper.Monitor.OpenInfo( "failing_db_setup_does_not_execute_SettleContent (1/2)" ) )
                 {
-                    var con = ctx[TestHelper.GetConnectionString()];
-                    var lastSetup = (DateTime)con.ExecuteScalar( new SqlCommand( "select LastStartDate from CKCore.tSetupMemory where SurrogateId=0" ) );
-                    var times = BuggyPackage.ReadSettleContentInfo( con );
-                    if( times.Count > 0 ) times[0].SetupTime.Should().BeBefore( lastSetup );
+                    File.WriteAllText( _configFile, @"<Error ErrorStep=""Install"" />" );
+                    SharedEngine.Reset();
+                    SharedEngine.EngineResult.Status.Should().Be( RunStatus.Failed );
+                    using( var ctx = new SqlStandardCallContext( TestHelper.Monitor ) )
+                    {
+                        var con = ctx[TestHelper.GetConnectionString()];
+                        var lastSetup = (DateTime)con.ExecuteScalar( new SqlCommand( "select LastStartDate from CKCore.tSetupMemory where SurrogateId=0" ) );
+                        var times = BuggyPackage.ReadSettleContentInfo( con );
+                        if( times.Count > 0 ) times[0].SetupTime.Should().BeBefore( lastSetup );
+                    }
+                }
+                // Removed config file and runs a new setup.
+                File.Delete( _configFile );
+
+                using( TestHelper.Monitor.OpenInfo( "failing_db_setup_does_not_execute_SettleContent (2/2)" ) )
+                {
+                    SharedEngine.Reset();
+                    var p = SharedEngine.Map.StObjs.Obtain<BuggyPackage>();
+                    using( var ctx = new SqlStandardCallContext( TestHelper.Monitor ) )
+                    {
+                        var lastSetup = (DateTime)ctx[p.Database].ExecuteScalar( new SqlCommand( "select LastStartDate from CKCore.tSetupMemory where SurrogateId=0" ) );
+                        var times = p.ReadSettleContentInfo( ctx );
+                        times[0].SetupTime.Should().Be( lastSetup );
+                    }
                 }
             }
-            // Removed config file and runs a new setup.
-            File.Delete( _configFile );
-
-            using( TestHelper.Monitor.OpenInfo( "failing_db_setup_does_not_execute_SettleContent (2/2)" ) )
+            finally
             {
-                SharedEngine.Reset();
-                var p = SharedEngine.Map.StObjs.Obtain<BuggyPackage>();
-                using( var ctx = new SqlStandardCallContext( TestHelper.Monitor ) )
-                {
-                    var lastSetup = (DateTime)ctx[p.Database].ExecuteScalar( new SqlCommand( "select LastStartDate from CKCore.tSetupMemory where SurrogateId=0" ) );
-                    var times = p.ReadSettleContentInfo( ctx );
-                    times[0].SetupTime.Should().Be( lastSetup );
-                }
+                if( File.Exists( _configFile ) ) File.Delete( _configFile );
             }
         }
     }
